requirements.txt
docopt==0.6.2
Pillow==4.0.0